make possible to build with c++0x

+++ vcl/unx/gtk/a11y/atkwindow.cxx
@@ -296,13 +296,13 @@ ooo_window_wrapper_get_type (void)
 
         static const GTypeInfo typeInfo =
         {
-            type_query.class_size,
+            static_cast<guint16>(type_query.class_size),
             (GBaseInitFunc) NULL,
             (GBaseFinalizeFunc) NULL,
             (GClassInitFunc) ooo_window_wrapper_class_init,
             (GClassFinalizeFunc) NULL,
             NULL,
-            type_query.instance_size,
+            static_cast<guint16>(type_query.instance_size),
             0,
             (GInstanceInitFunc) NULL,
             NULL
